uil_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.19
django-appconf==1.1.0
django-celery-email==3.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-filter==25.1
djoser==2.3.1
# habanero==1.2.6
httpx>=0.24
pandas>=2.0.0
openpyxl>=3.1.0
uvicorn[standard]==0.35.0
# daphne==4.1.2
gunicorn==23.0.0
drf-yasg>=1.21.7
# channels==4.2.2
# channels_redis==4.2.1
tenacity==9.1.2
Faker==37.1.0 
factory_boy==3.3.3
idna==3.10
kombu==5.4.2
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==24.2
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
# redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
# django_redis==5.4.0
pdf2image==1.17.0
Pillow==11.2.1
rapidfuzz== 3.13.0
